requirements.txt
flask
datacommons-mcp
mcp
google-generativeai
